+++ R/extend.R
@@ -3,32 +3,29 @@ if ( !isGeneric("extend") ) {
     standardGeneric("extend"))
 }
 
-#' extend satellite object
+#' Extend a Satellite object
 #'
 #' @description
-#' The function is a wrapper around the \code{\link{extend}} function to 
-#' easily extend a Satellite object by an \code{\link{extent}} object.
+#' The function is a wrapper around \code{\link[raster]{extend}} to easily 
+#' extend a Satellite object to a larger spatial extent.
 #'
 #' @param x Satellite object.
-#' @param y \code{\link{extent}} object. 
-#' @param subset Logical; if \code{TRUE} (default), all layers but the extendped 
-#' ones are being dropped; if \code{FALSE}, extendped layers are appended to the 
+#' @param y Target \code{Extent}, see \code{\link[raster]{extent}}. 
+#' @param subset Logical. If \code{TRUE} (default), all layers but the extended 
+#' ones are being dropped, else the extended layers are appended to the initial
 #' Satellite object.
+#' @param value Fill value assigned to new cells passed to 
+#' \code{\link[raster]{extend}}, defaults to \code{NA}.
 #'
-#' @return A Satellite object consisting of extended layers only. If 
-#' \code{subset = FALSE}, a Satellite object with the extendped layers appended.
+#' @return A Satellite object consisting of extended layers only or, if 
+#' \code{subset = FALSE}, a Satellite object with the extended layers appended.
 #' 
 #' @export extend
 #' 
 #' @name extend
 #' @aliases extend,Satellite-method
 #'
-#' @details extends layers of a Satellite object to the size of a given 
-#' \code{raster::extent} object.
-#' 
-#' @references Please refer to the respective functions for references.
-#'  
-#' @seealso This function is a wrapper for \code{raster::extend}.
+#' @seealso This function is a wrapper around \code{\link[raster]{extend}}.
 #'
 #' @examples
 #' \dontrun{
@@ -38,7 +35,7 @@ if ( !isGeneric("extend") ) {
 #' sat <- satellite(files)
 #'
 #' ## geographic extent of georg-gassmann-stadium (utm 32-n)
-#' ext_ggs <- raster::extent(484015, 484143, 5627835, 5628020)
+#' ext_ggs <- raster::extent(482606.4, 482781.4, 5627239, 5627489)
 #' 
 #' ## extend satellite object by specified extent
 #' sat_ggs <- extend(sat, ext_ggs)
@@ -48,14 +45,15 @@ if ( !isGeneric("extend") ) {
 #' }
 setMethod("extend", 
           signature(x = "Satellite"), 
-          function(x, y, subset = TRUE) {
+          function(x, y, subset = TRUE, value = NA) {
             rad_bands <- getSatBCDE(x)
             for (bcde_rad in rad_bands) {
-              ref <- extend(getSatDataLayer(x, bcde_rad), y)
-              # keep all metadata except for file path since extendped 
+
+              ref <- raster::extend(getSatDataLayer(x, bcde_rad), y, value)
+              # keep all metadata except for file path since extended 
               # layers are in memory and set calib column flag.
               meta_param <- getSatMeta(x, bcde_rad)
-              meta_param$CALIB <- "extendped"
+              meta_param$CALIB <- "extended"
               meta_param$FILE <- NULL
               
               info <- sys.calls()[[1]]
@@ -67,7 +65,7 @@ setMethod("extend",
             }
             
             if(subset == TRUE){
-              x <- subset(x, cid = "extendped")
+              x <- subset(x, cid = "extended")
               #reset LNBR (dirty hack)
               x@meta$LNBR <- rep(1:nrow(x@meta))
               x@meta$CALIB <- "SC"